+++ CHANGES.rst
@@ -49,7 +49,10 @@ New Features
 - ``photutils.segmentation``
 
   - Added a modified, significantly faster, ``SourceCatalog`` class.
-    [#1170]
+    [#1170, #1188]
+
+  - Added ``circular_aperture`` and ``circular_phometry`` methods to the
+    ``SourceCatalog`` class. [#1188]
 
   - Added a ``bbox`` attribute to ``SegmentationImage``. [#1187]